# api/openapi.json
{
  "components": {
    "schemas": {
...
      "CategoryOut": {
        "properties": {
          "icon": {
            "title": "Icon",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "owner_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "scope": {
            "title": "Scope",
            "type": "string"
          }
        },
        "required": [
          "id",
          "key",
          "label",
          "icon",
          "scope"
        ],
        "title": "CategoryOut",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
      "ExpenseIn": {
        "properties": {
          "amount_minor": {
            "exclusiveMinimum": 0.0,
            "title": "Amount Minor",
            "type": "integer"
          },
          "category_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Id"
          },
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "description": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Description",
            "type": "string"
          },
          "id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "notes": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "payer_id": {
            "format": "uuid",
            "title": "Payer Id",
            "type": "string"
          },
          "split": {
            "$ref": "#/components/schemas/SplitIn"
          },
          "version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "required": [
          "description",
          "amount_minor",
          "payer_id",
          "date",
          "split"
        ],
        "title": "ExpenseIn",
        "type": "object"
      },
      "ExpenseOut": {
        "properties": {
          "acting_admin_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Acting Admin Id"
          },
          "amount_minor": {
            "title": "Amount Minor",
            "type": "integer"
          },
          "category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CategoryOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "created_by": {
            "$ref": "#/components/schemas/UserBrief"
          },
          "currency": {
            "title": "Currency",
            "type": "string"
          },
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "deleted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted At"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "group_id": {
            "format": "uuid",
            "title": "Group Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "import_batch_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Import Batch Id"
          },
          "my_net_minor": {
            "title": "My Net Minor",
            "type": "integer"
          },
          "my_share_minor": {
            "title": "My Share Minor",
            "type": "integer"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "payer": {
            "$ref": "#/components/schemas/UserBrief"
          },
          "split_method": {
            "title": "Split Method",
            "type": "string"
          },
          "splits": {
            "items": {
              "$ref": "#/components/schemas/SplitOut"
            },
            "title": "Splits",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "group_id",
          "description",
          "amount_minor",
          "currency",
          "payer",
          "date",
          "category",
          "notes",
          "split_method",
          "splits",
          "my_share_minor",
          "my_net_minor",
          "created_by",
          "acting_admin_id",
          "import_batch_id",
          "created_at",
          "updated_at",
          "deleted_at",
          "version"
        ],
        "title": "ExpenseOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HistoryEntry": {
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "admin": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserBrief"
              },
              {
                "type": "null"
              }
            ]
          },
          "at": {
            "format": "date-time",
            "title": "At",
            "type": "string"
          },
          "diff": {
            "additionalProperties": true,
            "title": "Diff",
            "type": "object"
          },
          "person": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserBrief"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "action",
          "person",
          "admin",
          "at",
          "diff"
        ],
        "title": "HistoryEntry",
        "type": "object"
      },
...
      "SplitIn": {
        "properties": {
          "inputs": {
            "items": {
              "$ref": "#/components/schemas/SplitInputIn"
            },
            "title": "Inputs",
            "type": "array"
          },
          "method": {
            "title": "Method",
            "type": "string"
          }
        },
        "required": [
          "method",
          "inputs"
        ],
        "title": "SplitIn",
        "type": "object"
      },
      "SplitInputIn": {
        "properties": {
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          },
          "value": {
            "anyOf": [
              {
                "examples": [
                  "33.34"
                ],
                "pattern": "^-?\\d+(\\.\\d+)?$",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          }
        },
        "required": [
          "user_id"
        ],
        "title": "SplitInputIn",
        "type": "object"
      },
...
      "SplitOut": {
        "properties": {
          "input_value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Value"
          },
          "share_minor": {
            "title": "Share Minor",
            "type": "integer"
          },
          "status": {
            "enum": [
              "paid",
              "owes",
              "settled"
            ],
            "title": "Status",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/UserBrief"
          }
        },
        "required": [
          "user",
          "share_minor",
          "input_value",
          "status"
        ],
        "title": "SplitOut",
        "type": "object"
      },
...
      "UserBrief": {
        "description": "A person as shown inside groups, splits and balances.",
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_placeholder": {
            "default": false,
            "title": "Is Placeholder",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "avatar_url"
        ],
        "title": "UserBrief",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Splitkit API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/expenses/{expense_id}": {
      "delete": {
        "description": "Soft delete. The expense stops counting towards balances and can be restored.",
        "operationId": "delete_expense",
        "parameters": [
          {
            "in": "path",
            "name": "expense_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Expense Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Expense",
        "tags": [
          "expenses"
        ]
      },
      "get": {
        "operationId": "get_expense",
        "parameters": [
          {
            "in": "path",
            "name": "expense_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Expense Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Expense",
        "tags": [
          "expenses"
        ]
      },
      "put": {
        "operationId": "update_expense",
        "parameters": [
          {
            "in": "path",
            "name": "expense_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Expense Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Expense",
        "tags": [
          "expenses"
        ]
      }
    },
    "/api/v1/expenses/{expense_id}/history": {
      "get": {
        "description": "Who added and changed this expense, oldest first (\"Edited by Rahul · ₹1,650 → ₹1,800\").",
        "operationId": "expense_history",
        "parameters": [
          {
            "in": "path",
            "name": "expense_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Expense Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HistoryEntry"
                  },
                  "title": "Response Expense History",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Expense History",
        "tags": [
          "expenses"
        ]
      }
    },
    "/api/v1/expenses/{expense_id}/restore": {
      "post": {
        "operationId": "restore_expense",
        "parameters": [
          {
            "in": "path",
            "name": "expense_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Expense Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Restore Expense",
        "tags": [
          "expenses"
        ]
      }
    },
...
    "/api/v1/groups/{group_id}/expenses": {
      "post": {
        "operationId": "create_expense",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Group Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExpenseIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Expense",
        "tags": [
          "expenses"
        ]
      }
    },
...
  }
}